backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
mysql-connector-python
python-dotenv
Flask-Cors
joblib
Flask-JWT-Extended
pandas
scikit-learn